requirements.txt
streamlit
google-generativeai
SpeechRecognition
pyttsx3